# webapp/contracts/openapi/bundled.yaml
openapi: 3.1.0
info:
  title: BizSim Edu API
  description: 商域 - 商业模拟教育平台后端 API
  version: 1.0.0
paths:
...
  /api/v1/teaching-groups:
    post:
      tags:
      - 体验营
      summary: Create Teaching Group
      operationId: create_teaching_group_api_v1_teaching_groups_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TeachingGroupCreate'
        required: true
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse_TeachingGroupOut_'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - OAuth2PasswordBearer: []
...
  /api/v1/teaching-groups/{group_id}/groups:
    get:
      tags:
      - 营团分组
      summary: List Groups
      operationId: list_groups_api_v1_teaching_groups__group_id__groups_get
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: group_id
        in: path
        required: true
        schema:
          type: integer
          title: Group Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse_List_CampGroupDetail__'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    post:
      tags:
      - 营团分组
      summary: Create Group
      operationId: create_group_api_v1_teaching_groups__group_id__groups_post
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: group_id
        in: path
        required: true
        schema:
          type: integer
          title: Group Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CampGroupCreate'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse_CampGroupOut_'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/teaching-groups/{group_id}/groups/{camp_group_id}:
    patch:
      tags:
      - 营团分组
      summary: Update Group
      operationId: update_group_api_v1_teaching_groups__group_id__groups__camp_group_id__patch
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: group_id
        in: path
        required: true
        schema:
          type: integer
          title: Group Id
      - name: camp_group_id
        in: path
        required: true
        schema:
          type: integer
          title: Camp Group Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CampGroupUpdate'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse_CampGroupOut_'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    delete:
      tags:
      - 营团分组
      summary: Delete Group
      operationId: delete_group_api_v1_teaching_groups__group_id__groups__camp_group_id__delete
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: group_id
        in: path
        required: true
        schema:
          type: integer
          title: Group Id
      - name: camp_group_id
        in: path
        required: true
        schema:
          type: integer
          title: Camp Group Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/teaching-groups/{group_id}/groups/auto-generate:
    post:
      tags:
      - 营团分组
      summary: Auto Generate Groups
      operationId: auto_generate_groups_api_v1_teaching_groups__group_id__groups_auto_generate_post
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: group_id
        in: path
        required: true
        schema:
          type: integer
          title: Group Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AutoGroupRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/teaching-groups/{group_id}/groups/{camp_group_id}/members:
    post:
      tags:
      - 营团分组
      summary: Add Member To Group
      operationId: add_member_to_group_api_v1_teaching_groups__group_id__groups__camp_group_id__members_post
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: group_id
        in: path
        required: true
        schema:
          type: integer
          title: Group Id
      - name: camp_group_id
        in: path
        required: true
        schema:
          type: integer
          title: Camp Group Id
      - name: user_id
        in: query
        required: true
        schema:
          type: integer
          title: User Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/teaching-groups/{group_id}/groups/{camp_group_id}/members/{user_id}:
    delete:
      tags:
      - 营团分组
      summary: Remove Member From Group
      operationId: remove_member_from_group_api_v1_teaching_groups__group_id__groups__camp_group_id__members__user_id__delete
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: group_id
        in: path
        required: true
        schema:
          type: integer
          title: Group Id
      - name: camp_group_id
        in: path
        required: true
        schema:
          type: integer
          title: Camp Group Id
      - name: user_id
        in: path
        required: true
        schema:
          type: integer
          title: User Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/teaching-groups/{group_id}/groups/{camp_group_id}/members/{user_id}/role:
    patch:
      tags:
      - 营团分组
      summary: Set Member Role
      operationId: set_member_role_api_v1_teaching_groups__group_id__groups__camp_group_id__members__user_id__role_patch
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: group_id
        in: path
        required: true
        schema:
          type: integer
          title: Group Id
      - name: camp_group_id
        in: path
        required: true
        schema:
          type: integer
          title: Camp Group Id
      - name: user_id
        in: path
        required: true
        schema:
          type: integer
          title: User Id
      - name: role
        in: query
        required: true
        schema:
          type: string
          title: Role
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    ApiResponse_CampGroupOut_:
      properties:
        success:
          type: boolean
          title: Success
          default: true
        code:
          type: integer
          title: Code
          default: 0
        message:
          type: string
          title: Message
          default: success
        data:
          anyOf:
          - $ref: '#/components/schemas/CampGroupOut'
          - type: 'null'
      type: object
      title: ApiResponse[CampGroupOut]
...
    ApiResponse_List_CampGroupDetail__:
      properties:
        success:
          type: boolean
          title: Success
          default: true
        code:
          type: integer
          title: Code
          default: 0
        message:
          type: string
          title: Message
          default: success
        data:
          anyOf:
          - items:
              $ref: '#/components/schemas/CampGroupDetail'
            type: array
          - type: 'null'
          title: Data
      type: object
      title: ApiResponse[List[CampGroupDetail]]
...
    ApiResponse_TeachingGroupOut_:
      properties:
        success:
          type: boolean
          title: Success
          default: true
        code:
          type: integer
          title: Code
          default: 0
        message:
          type: string
          title: Message
          default: success
        data:
          anyOf:
          - $ref: '#/components/schemas/TeachingGroupOut'
          - type: 'null'
      type: object
      title: ApiResponse[TeachingGroupOut]
...
    AutoGroupRequest:
      properties:
        group_size:
          type: integer
          maximum: 20.0
          minimum: 2.0
          title: Group Size
        method:
          type: string
          pattern: ^(random|join_order)$
          title: Method
          default: random
      type: object
      required:
      - group_size
      title: AutoGroupRequest
...
    CampGroupCreate:
      properties:
        name:
          type: string
          maxLength: 64
          minLength: 1
          title: Name
        color:
          anyOf:
          - type: string
          - type: 'null'
          title: Color
      type: object
      required:
      - name
      title: CampGroupCreate
    CampGroupDetail:
      properties:
        id:
          type: integer
          title: Id
        teaching_group_id:
          type: integer
          title: Teaching Group Id
        name:
          type: string
          title: Name
        color:
          anyOf:
          - type: string
          - type: 'null'
          title: Color
        member_count:
          type: integer
          title: Member Count
          default: 0
        created_at:
          type: string
          format: date-time
          title: Created At
        members:
          items:
            $ref: '#/components/schemas/CampGroupMemberOut'
          type: array
          title: Members
          default: []
      type: object
      required:
      - id
      - teaching_group_id
      - name
      - created_at
      title: CampGroupDetail
    CampGroupMemberOut:
      properties:
        user_id:
          type: integer
          title: User Id
        username:
          type: string
          title: Username
        role:
          type: string
          title: Role
          default: member
        joined_at:
          type: string
          format: date-time
          title: Joined At
      type: object
      required:
      - user_id
      - username
      - joined_at
      title: CampGroupMemberOut
    CampGroupOut:
      properties:
        id:
          type: integer
          title: Id
        teaching_group_id:
          type: integer
          title: Teaching Group Id
        name:
          type: string
          title: Name
        color:
          anyOf:
          - type: string
          - type: 'null'
          title: Color
        member_count:
          type: integer
          title: Member Count
          default: 0
        created_at:
          type: string
          format: date-time
          title: Created At
      type: object
      required:
      - id
      - teaching_group_id
      - name
      - created_at
      title: CampGroupOut
    CampGroupUpdate:
      properties:
        name:
          anyOf:
          - type: string
            maxLength: 64
            minLength: 1
          - type: 'null'
          title: Name
        color:
          anyOf:
          - type: string
          - type: 'null'
          title: Color
      type: object
      title: CampGroupUpdate
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    TeachingGroupCreate:
      properties:
        name:
          type: string
          maxLength: 120
          minLength: 1
          title: Name
        description:
          anyOf:
          - type: string
          - type: 'null'
          title: Description
        camp_start_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Camp Start At
        camp_end_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Camp End At
      type: object
      required:
      - name
      title: TeachingGroupCreate
...
    TeachingGroupOut:
      properties:
        id:
          type: integer
          title: Id
        name:
          type: string
          title: Name
        description:
          anyOf:
          - type: string
          - type: 'null'
          title: Description
        invite_code:
          type: string
          title: Invite Code
        teacher_user_id:
          type: integer
          title: Teacher User Id
        teacher_username:
          anyOf:
          - type: string
          - type: 'null'
          title: Teacher Username
        camp_start_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Camp Start At
        camp_end_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Camp End At
        status:
          type: string
          title: Status
        created_at:
          type: string
          format: date-time
          title: Created At
        member_count:
          type: integer
          title: Member Count
          default: 0
        event_count:
          type: integer
          title: Event Count
          default: 0
      type: object
      required:
      - id
      - name
      - invite_code
      - teacher_user_id
      - status
      - created_at
      title: TeachingGroupOut
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
  securitySchemes:
    OAuth2PasswordBearer:
      type: oauth2
      flows:
        password:
          scopes: {}
          tokenUrl: /api/v1/auth/login